Move pydantic models out to their own module
